process function complete

+++ include/process/chs_ct.h
@@ -2,6 +2,8 @@
 #ifndef __CHS_CT_H__
 #define __CHS_CT_H__
 
+#include <stdio.h>
+
 typedef struct city{
 	unsigned int ct_id;		//city id
 	char ct_nm[128];	//city name 
@@ -11,7 +13,7 @@ typedef struct city{
 /*
 *according the value of ct_id,converting id to city name stored to ct_nm, the city name length is limited by the argument @ct_len
 */
-extern int id_to_city(unsigned int ct_id, char* ct_nm, unsigned int ct_len);
+extern int id_to_city(unsigned int ct_id, char* ct_nm, ssize_t ct_len);
 
 /*
 *loading city info list from the file pointed to by ct_fnm */
