+++ include/process/chs_pv.h
@@ -1,6 +1,7 @@
 /*src/process/chs_pv.h*/
 #ifndef __CHS_PV_H__
 #define __CHS_PV_H__
+#include <stdio.h>
 
 typedef struct province{
 	unsigned int pv_id;
@@ -11,7 +12,7 @@ typedef struct province{
 /*
 *according the value of pv_id,converting id to province name stored to pv_nm, the province name length is limited by the argument @pv_len
 */
-extern int id_to_province(unsigned int pv_id, char* pv_nm, unsigned int pv_len);
+extern int id_to_province(unsigned int pv_id, char* pv_nm, ssize_t pv_len);
 
 /*
 *loading province info list from the file pointed to by pv_fnm */
